Rename JS controller
