+++ app/javascript/controllers/dropzone_controller.js
