+++ app/javascript/controllers/index.js
@@ -28,8 +28,8 @@ application.register("form", FormController)
 import MessageDraftsController from "./message_drafts_controller"
 application.register("message-drafts", MessageDraftsController)
 
-import FsMessageDraftsController from "./fs_message_drafts_controller"
-application.register("fs-message-drafts", FsMessageDraftsController)
+import DropzoneController from "./dropzone_controller"
+application.register("dropzone", DropzoneController)
 
 import TriStateCheckboxController from "./tri_state_checkbox_controller"
 application.register("tri-state-checkbox", TriStateCheckboxController)
