+++ app/views/fs/message_drafts/_new_form.html.erb
@@ -5,7 +5,7 @@
         <label for="cover-photo" class="block text-sm font-medium leading-6 text-gray-900">XML obsah správ</label>
         <div id="dropzone" class="mt-2 flex justify-center rounded-lg border border-dashed border-gray-900/25 px-6 py-10">
           <%= content_tag(:div, {
-            "data-controller": "fs-message-drafts",
+            "data-controller": "dropzone",
             class: "text-center mt-6"
           }) do %>
             <%= render Common::IconComponent.new("document-text", classes: "mx-auto h-12 w-12 text-gray-300") %>
